requirements.txt
absl-py==0.7.1
astor==0.8.0
backcall==0.1.0
decorator==4.4.0
gast==0.2.2
graphviz==0.11
grpcio==1.21.1
h5py==2.9.0
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
numpy==1.16.4
opencv-python==4.1.0.25
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
Pygments==2.4.2
PyYAML==5.1.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
torch==1.1.0
torchvision==0.3.0
torchviz==0.0.1
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.4
